doctor-patient-portal/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
Pillow
drf_yasg